Update entrypoint to application (#1)
* Update entrypoint-to-application
* Update Dockerfile & entrypoint, now:
good coop between Dockerfile and entrypoint

+++ entrypoint.sh
@@ -1,22 +1,64 @@
 #!/usr/bin/env bash
 
-set -e
+set -eo pipefail
 
-# Empty apache2 logs directory
-if [ -d /var/log/apache2 ]; then
-    echo "cleanup apache2 logs"
-    rm -rf /var/log/apache2/*.log
+
+INSTALL_SCRIPT=resetdevelopment.sh
+
+# if not additional instructions are send
+if [ "$1" = '/usr/bin/supervisord' -a "$2" = '-c' -a "$3" = '/etc/supervisor/supervisord.conf' ]; then
+
+
+    # Empty apache2 logs directory
+    if [ -d /var/log/apache2 ]; then
+        echo "cleanup apache2 logs"
+        rm -rf /var/log/apache2/*.log
+    fi
+
+    # Empty supervisor logs directory
+    if [ -d /var/log/supervisor ]; then
+        echo "cleanup supervisord logs"
+        rm -rf /var/log/supervisor/*.log
+    fi
+
+
+
+    # Empty supervisor logs directory
+    if [ ! -d "${PROJECT_PATH}" ]; then
+
+        git clone -b "${PROJECT_BRANCH}" --single-branch "https://${GITHUB_OAUTH}:x-oauth-basic@${PROJECT_REPOSITORY}" "${PROJECT_PATH}"
+
+        cd "${PROJECT_PATH}"
+
+        git remote set-url origin "https://${PROJECT_REPOSITORY}"
+
+        composer config -g "github-oauth.github.com" "${GITHUB_OAUTH}"
+
+
+        # Empty supervisor logs directory
+        if [ -f "${PROJECT_PATH}/${INSTALL_SCRIPT}" ]; then
+            /bin/sh -c "${PROJECT_PATH}/${INSTALL_SCRIPT}"
+        fi
+
+    else
+
+        # prevent unwanted deleting
+        echo ""
+        echo ""
+        echo ""
+        echo "directory already exists!"
+        echo ""
+        echo ""
+        echo ""
+        echo ""
+        echo ""
+        echo ""
+
+    fi
+
+    # Start services
+    echo "Start magentostack"
+
 fi
 
-# Empty supervisor logs directory
-if [ -d /var/log/supervisor ]; then
-    echo "cleanup supervisord logs"
-    rm -rf /var/log/supervisor/*.log
-fi
-
-composer config -g "github-oauth.github.com" "$GITHUB_OAUTH"
-
-# Start services
-echo "Start magentostack"
-
-/usr/bin/supervisord -c /etc/supervisor/supervisord.conf
+exec "$@"
